requirements.txt
beautifulsoup4
requests
python-telegram-bot
asyncio
